+++ packages/app/src/index.ts
@@ -1,4 +1,5 @@
 export * from './runtime'
 export * from './oplog'
+export * from './repairindex'
 export * from './http'
 export { buildServerDeps, startServer } from './server'